backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
pydantic-settings
python-dotenv
firebase-admin
sentry-sdk[fastapi]
python-slowapi
email-validator # Required by pydantic's EmailStr

# For testing (optional but recommended)
# pytest
# httpx
# pytest-asyncio